+++ src/violet/data/InitOpts.hpp
@@ -1,0 +1,13 @@
+#pragma once
+
+#include <string>
+
+namespace violet {
+
+struct InitOpts {
+    std::string folder;
+
+    std::string prefix = "/";
+};
+
+}