+++ src/violet/init/InitModule.hpp
@@ -1,0 +1,8 @@
+#pragma once
+
+#include "violet/data/InitOpts.hpp"
+namespace violet {
+
+extern int initMain(const InitOpts& opts);
+
+}